backend/requirements.txt
pip install \
fastapi \
uvicorn \
python-multipart \
pydantic \
langchain \
langchain-community \
sentence-transformers \
faiss-cpu \
pypdf \
pandas \
python-dotenv \
transformers \
torch
